extract class handlers
