+++ lib/handlers/classHandlers.js
@@ -1,4 +1,86 @@
 var tokenUtils = require('../util/tokenUtils');
+var handlerUtils = require('../util/handlerUtils');
+
+/**
+ * Handles class declarations:
+ *   classy <ClassName>
+ *
+ * Produces:
+ *   class <ClassName> {
+ */
+function handleClass(parseContext) {
+    tokenUtils.expectToken('classy', parseContext);
+  
+    // consume: classy
+    parseContext.tokens.shift();
+  
+    var tokens = parseContext.tokens;
+  
+    /**
+     * We're in a class expression without extension
+     */
+    if(tokens.length < 1)
+    {
+      return "class {\n";
+    }
+  
+    var className = parseContext.tokens.shift();
+    var statement = "class " + className;
+  
+    /**
+     * Check if there's a grows ParentClassName
+     */
+    if(tokens.length > 0)
+    {
+      var nextToken = tokens[0];
+      if(nextToken == 'grows') {
+        statement += " " + handleGrows(parseContext);
+      }
+    }
+  
+    statement += "{\n";
+    return statement;
+  }
+  
+  /**
+   * Handles constructor declarations:
+   * maker <args>
+   *
+   * Produces:
+   *  constructor <args> {
+   */
+  function handleMaker(parseContext) {
+    tokenUtils.expectToken('maker', parseContext);
+  
+    var tokens = parseContext.tokens;
+  
+    // consume: maker
+    tokens.shift();
+  
+    return  'constructor' + handlerUtils.declareArguments(parseContext);
+  }
+  
+  /**
+   * Handles class extension declarations:
+   * grows <ClassName>
+   *
+   * Produces:
+   *  extends <ClassName>
+   */
+  function handleGrows(parseContext) {
+    tokenUtils.expectToken('grows', parseContext);
+  
+    var tokens = parseContext.tokens;
+  
+    // consume: grows
+    tokens.shift();
+  
+    if (tokens.length < 1) {
+      throw new SyntaxError(`Expected ClassName but got nothing. Allowed construct: grows [ClassName]. ${parseInfo(parseContext)}`);
+    }
+  
+    return 'extends ' + tokens.shift();
+  }
 
 /**
  * Handles getter method declarations:
@@ -64,5 +146,5 @@ function handleSit(parseContext) {
 }
 
 module.exports = {
-    handleGit, handleSit
+    handleGit, handleSit , handleClass, handleGrows, handleMaker
 }
